Move framebuf attachment into central code.
It doesn't look like this will be different, and there's a bug in it.
